requirements.txt
PyPDF2
pillow
streamlit 